Asyncify outermost logic in Blobimport::import
Summary:
In the previous diff we asyncified the signature of Blobimport::import,
but the body remained an old-style future with a compat and await at
the end.

This diff asyncifies the outermost logic from within the function,
slightly improving readability and removing one layer of clones
to cut down on heap allocations. The derivation of `max_rev` still
currently uses old-style streams and futures.

Reviewed By: StanislavGlebik

Differential Revision: D20861230

fbshipit-source-id: 1b462f17581c764e77a0a0a163c86ffa894df742