Asyncify the max_rev code
Summary:
This wound up being a little tricky, because
that `async move` blocks capture any data used,
and most of the fields of the `Blobimport` struct
are values rather than refs.

The easiest solution that I came up with, which looks
a little weird but works better than anything else
I tried, is to just inject a little block of code
(which I commented so it will hopefully be clear to
future readers) taking refs of anything that we need
to use in an async block but also have available later.

In the process, we are able to strip out a layer of
clones, which should improve efficiency a bit.

Reviewed By: StanislavGlebik

Differential Revision: D20862358

fbshipit-source-id: 186bf9939b9496c432ff0d9a01e602da47f4b5d4