disable reading blob from union store
Summary:
This code isn't really working and is giving out corrupted file content. Let's remove it for reading blobs.

We still cannot remove the C++ mercurial code entirely as we rely on it to import manifests. I tested on switching to the Rust manifest implementation but it wasn't working due to new pack file scanning.

Reviewed By: chadaustin

Differential Revision: D20938001

fbshipit-source-id: 9cc7b1ed9cd8228afd6fc5374cda001d1309a379